athene_setup: Check clean build

+++ athene_setup.sh
@@ -1,1 +1,7 @@
-rm -rf out/target/product/athene
+#!/bin/bash
+printf "Clean build: "
+read choice
+if [ $choice == 'y' ]
+then
+    rm -rf out/target/product/athene
+fi
